+++ CHANGELOG.md
@@ -310,6 +310,8 @@
 - Only show the edit transaction link when the transaction can be edited
 - Only show the add transaction button when the activity can be edited
 - Updated planned disbursement policy
+- Only show the edit planned disbursement link then it the planned disbursement
+  can be edited
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-16...HEAD
 [release-16]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-15...release-16